faser ut metode for reseting av steg fra gammel StegService. Tar i bruk ny StegService sin metode i stedet
